+++ core/src/main/java/com/github/kaktushose/jda/commands/annotations/i18n/Bundle.java
@@ -1,5 +1,7 @@
 package com.github.kaktushose.jda.commands.annotations.i18n;
 
+import com.github.kaktushose.jda.commands.message.i18n.I18n;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -8,7 +10,7 @@ import java.lang.annotation.Target;
 /// The bundle annotation states which localization bundle is to be used
 /// in a certain scope.
 ///
-/// @see com.github.kaktushose.jda.commands.i18n.I18n
+/// @see I18n
 @Target({ElementType.PACKAGE, ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Bundle {
