+++ core/src/main/java/com/github/kaktushose/jda/commands/dispatching/reply/dynamic/ButtonComponent.java
@@ -2,7 +2,7 @@ package com.github.kaktushose.jda.commands.dispatching.reply.dynamic;
 
 import com.github.kaktushose.jda.commands.definitions.interactions.component.ButtonDefinition;
 import com.github.kaktushose.jda.commands.dispatching.reply.Component;
-import com.github.kaktushose.jda.commands.i18n.I18n;
+import com.github.kaktushose.jda.commands.message.i18n.I18n;
 import net.dv8tion.jda.api.entities.emoji.Emoji;
 import net.dv8tion.jda.api.interactions.components.buttons.Button;
 import net.dv8tion.jda.api.interactions.components.buttons.ButtonStyle;
