+++ core/src/main/java/com/github/kaktushose/jda/commands/exceptions/ConfigurationException.java
@@ -1,7 +1,7 @@
 package com.github.kaktushose.jda.commands.exceptions;
 
 import com.github.kaktushose.jda.commands.exceptions.internal.JDACException;
-import com.github.kaktushose.jda.commands.i18n.I18n;
+import com.github.kaktushose.jda.commands.message.i18n.I18n;
 
 /// Will be thrown if anything goes wrong while configuring jda-commands.
 public final class ConfigurationException extends JDACException {
