+++ core/src/main/java/com/github/kaktushose/jda/commands/exceptions/InternalException.java
@@ -1,7 +1,7 @@
 package com.github.kaktushose.jda.commands.exceptions;
 
 import com.github.kaktushose.jda.commands.exceptions.internal.JDACException;
-import com.github.kaktushose.jda.commands.i18n.I18n;
+import com.github.kaktushose.jda.commands.message.i18n.I18n;
 
 import java.util.Locale;
 import java.util.Map;
