+++ core/src/main/java/com/github/kaktushose/jda/commands/exceptions/InvalidDeclarationException.java
@@ -3,7 +3,7 @@ package com.github.kaktushose.jda.commands.exceptions;
 import com.github.kaktushose.jda.commands.definitions.description.MethodDescription;
 import com.github.kaktushose.jda.commands.definitions.description.ParameterDescription;
 import com.github.kaktushose.jda.commands.exceptions.internal.JDACException;
-import com.github.kaktushose.jda.commands.i18n.I18n;
+import com.github.kaktushose.jda.commands.message.i18n.I18n;
 
 import java.util.stream.Collectors;
