+++ core/src/main/java/module-info.java
@@ -68,7 +68,7 @@ module io.github.kaktushose.jda.commands.core {
     exports com.github.kaktushose.jda.commands.extension;
 
     // i18n/messages
-    exports com.github.kaktushose.jda.commands.i18n;
+    exports com.github.kaktushose.jda.commands.message.i18n;
     exports com.github.kaktushose.jda.commands.message;
 
     uses com.github.kaktushose.jda.commands.extension.Extension;
