+++ guice-extension/src/main/java/com/github/kaktushose/jda/commands/guice/internal/GuiceInteractionControllerInstantiator.java
@@ -2,7 +2,7 @@ package com.github.kaktushose.jda.commands.guice.internal;
 
 import com.github.kaktushose.jda.commands.annotations.interactions.Interaction;
 import com.github.kaktushose.jda.commands.dispatching.instance.InteractionControllerInstantiator;
-import com.github.kaktushose.jda.commands.i18n.I18n;
+import com.github.kaktushose.jda.commands.message.i18n.I18n;
 import com.google.inject.Injector;
 import net.dv8tion.jda.api.JDA;
 import org.jetbrains.annotations.ApiStatus;
